requirements.txt
torch>=2.0.0
torchvision>=0.15.0
numpy>=1.24.0
pandas>=2.0.0
certifi>=2023.0.0
pyyaml>=6.0

# Profiling (thesis 3.7.3: FLOPs via fvcore; thop optional fallback)
fvcore>=0.1.5
thop>=0.1.1
